Split z64.h more
