+++ include/z64dma.h
@@ -22,6 +22,18 @@ typedef struct {
     /* 0x0C */ uintptr_t romEnd;
 } DmaEntry;
 
+typedef struct {
+    /* 0x00 */ uintptr_t vromStart;
+    /* 0x04 */ uintptr_t vromEnd;
+} RomFile; // size = 0x8
+
+#define ROM_FILE(name) \
+    { (uintptr_t)_##name##SegmentRomStart, (uintptr_t)_##name##SegmentRomEnd }
+#define ROM_FILE_EMPTY(name) \
+    { (uintptr_t)_##name##SegmentRomStart, (uintptr_t)_##name##SegmentRomStart }
+#define ROM_FILE_UNSET \
+    { 0 }
+
 extern DmaEntry gDmaDataTable[];
 
 extern u32 gDmaMgrVerbose;
