+++ include/z64skybox.h
@@ -5,7 +5,7 @@
 #include "ultra64/gbi.h"
 
 #include "z64math.h"
-#include "z64scene.h" // for RomFile
+#include "z64dma.h" // for RomFile
 
 struct GameState;
 struct GraphicsContext;
