+++ include/z64state.h
@@ -1,0 +1,37 @@
+#ifndef Z64STATE_H
+#define Z64STATE_H
+
+#include "gfx.h"
+#include "padmgr.h"
+
+typedef struct GameAllocEntry {
+    /* 0x00 */ struct GameAllocEntry* next;
+    /* 0x04 */ struct GameAllocEntry* prev;
+    /* 0x08 */ u32 size;
+    /* 0x0C */ u32 unk_0C;
+} GameAllocEntry; // size = 0x10
+
+typedef struct {
+    /* 0x00 */ GameAllocEntry base;
+    /* 0x10 */ GameAllocEntry* head;
+} GameAlloc; // size = 0x14
+
+struct GameState;
+
+typedef void (*GameStateFunc)(struct GameState* gameState);
+
+typedef struct GameState {
+    /* 0x00 */ GraphicsContext* gfxCtx;
+    /* 0x04 */ GameStateFunc main;
+    /* 0x08 */ GameStateFunc destroy; // "cleanup"
+    /* 0x0C */ GameStateFunc init;
+    /* 0x10 */ u32 size;
+    /* 0x14 */ Input input[MAXCONTROLLERS];
+    /* 0x74 */ TwoHeadArena tha;
+    /* 0x84 */ GameAlloc alloc;
+    /* 0x98 */ u32 running;
+    /* 0x9C */ u32 frames;
+    /* 0xA0 */ u32 inPreNMIState;
+} GameState; // size = 0xA4
+
+#endif
